backend/pyproject.toml
[project]
name = "gonza-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pandas>=3.0.1",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "reportlab>=4.3.1",
    "django-filter>=25.2",
    "django-unfold>=0.85.0",
    "cuid2>=2.0.1",
    "python-dotenv>=1.2.2",
]
